Refine `fuzzy_score`
The diff looks larger than it maybe should.

- Give matches at the start 2 extra points
- Do not give penalty for consecutive matching characters (!)

The higher the score the unlikely we suggest it.  So we could as well
say the `score` is a "penalty".  For consecutive matches, we still payed
"1" (as in `pos - prev`); we remove here.

This is WIP as in this commit message `consectuv<TAB>` will not suggest
"consecutive".  ST was always able to fix simple typos as well.

Other than that this algorithm mostly suggests what ST suggests on
`alt-space` in the *same* order.