Do exclude the `initial_primer` from the word list
Usually the `initial_primer` is a best match.  Don't include it in the
word list so we don't have to skip over it initially.

Note that we always include the primer through `_matching()`. What we
don't want is: after suggesting the history item, first suggest the
primer as its probably the top match from the current view, and only
then show the next suggestion from the view.