Fix: Allow completing in the middle of a word
Following ST, you can ask for completion being inside of a word.  In
that case `primer` is the region from the start of the word to the
cursor.

For example, consider you have this

```
    if ch == "A":
```

and want to add another condition at the beginning, `|` denoting the
cursor:

```
		if sc|ch == "A":
```

Pressing `tab` should now for example expand to

```
		if score|ch == "A":
```