Prevent abbrevations from entering the word lists
Generally the indexer adds the word under the cursor to the index.
Although the local ("by view") list is temporary, it has the negative
effect that fantasy words get in the index and in our way.

For example, with `|` denoting the cursor:

```
if sc|ch == "A"
```

here the word "scch" enters the index which is also typically a "good"
match for our algorithm as it shares the prefix with our primer.

For the global everstore this is a no-brainer because the everstore
really fills up with short abbreviations.